adding restore deleted data and putting the get queries outside of the auth group

+++ capstone-admin/app/Http/Controllers/Government/NewsController.php
@@ -6,10 +6,63 @@ use App\Models\Government\News;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Validator;
-use Image, File;
+use Image, Throwable;
 
 class NewsController extends Controller
 {
+    /*
+    * get all trashed data
+    */
+    public function getAllTrashed(Request $request)
+    {
+        try {
+            $news_data = News::onlyTrashed()->get();
+            if ($news_data) {
+                return response()->json(["news" => $news_data]);
+            }
+            return response()->json(["msg" => $news_data, "status" => 404], 404);
+        } catch (\Throwable $err) {
+            //throw $th;
+            return response()->json(["res" => ["msg" => $err->getMessage(), "status" => 400]], 400);
+        }
+    }
+
+    /*
+    * restore trashed data
+    */
+    public function restore(Request $request)
+    {
+        // get the passed parameter id
+        // validate
+        $validator = Validator::make($request->all(), [
+            "id" => "required",
+        ]);
+
+        /*
+        * customizing the validation response
+        */
+        if ($validator->fails()) {
+            # send the second error message if exists otherwise send the first one
+            return response()->json([
+                "res" => [
+                    "msg" => $validator->messages()->all()[1] ?: $validator->messages()->all()[0],
+                    "status" => 400,
+                ]
+            ], 400);
+        }
+
+        $id = $request->input("id");
+        try {
+            News::withTrashed()->find($id)->restore();
+
+            # return all the trashed data
+            return response()->json(["news" => News::onlyTrashed()->get(), "res" => ["msg" => "successfully restore data", "status" => 200]], 200);
+        } catch (\Throwable $err) {
+            //throw $th;
+            return response()->json(["res" => ["msg" => $err->getMessage(), "status" => 400]], 400);
+        }
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -22,6 +75,9 @@ class NewsController extends Controller
         return inertia("Main");
     }
 
+    /*
+    * get all not trashed data
+    */
     public function getNews()
     {
         //
@@ -32,15 +88,32 @@ class NewsController extends Controller
     public function deleteOneNews(Request $request)
     {
         // get the passed parameter id
-        $request->validate(["id" => ["required"]]);
+        // validate
+        $validator = Validator::make($request->all(), [
+            "id" => "required",
+        ]);
+
+        /*
+        * customizing the validation response
+        */
+        if ($validator->fails()) {
+            # send the second error message if exists otherwise send the first one
+            return response()->json([
+                "res" => [
+                    "msg" => $validator->messages()->all()[1] ?: $validator->messages()->all()[0],
+                    "status" => 400,
+                ]
+            ], 400);
+        }
+
         $id = $request->input("id");
 
         try {
             News::findOrFail($id)->delete();
-            return response()->json(["news" => News::orderBy('created_at', 'desc')->get(), "res" => ["msg" => "Successfully deleted a news", "status" => 200]]);
+            return response()->json(["news" => News::orderBy('created_at', 'desc')->get(), "res" => ["msg" => "Successfully deleted a news", "status" => 200]], 200);
         } catch (\Throwable $th) {
             //throw $th;
-            return response()->json(["res" => ["msg" => $th, "status" => 400]]);
+            return response()->json(["res" => ["msg" => $th, "status" => 400]], 400);
         }
     }
 
@@ -67,13 +140,16 @@ class NewsController extends Controller
                     "msg" => $validator->messages()->all()[1] ?: $validator->messages()->all()[0],
                     "status" => 400,
                 ]
-            ]);
+            ], 400);
         }
 
         // get the passed parameter id
         $id = $request->input("id");
 
         try {
+            if (News::withTrashed()->find($id)->trashed()) {
+                return response()->json(["res" => ['msg' => "Please restore this data before you can edit this", 'status' => 400]], 400);
+            }
             $news = News::findOrFail($id);
             $news->title = $request->input("title");
             $news->description = $request->input("description");
@@ -141,11 +217,11 @@ class NewsController extends Controller
                 "res" => [
                     "msg" => "Successfully updated news",
                     "status" => 200
-                ]
+                ], 200
             ]);
         } catch (\Throwable $err) {
             //throw $th;
-            return response()->json(["res" => ["msg" => $err->getMessage(), "status" => 400]]);
+            return response()->json(["res" => ["msg" => $err->getMessage(), "status" => 400]], 400);
         }
     }
 
@@ -175,7 +251,7 @@ class NewsController extends Controller
                 "res" => [
                     "msg" => $validator->messages()->all()[1] ?: $validator->messages()->all()[0],
                     "status" => 400,
-                ]
+                ], 400
             ]);
         }
 
@@ -211,11 +287,11 @@ class NewsController extends Controller
                 "res" => [
                     "msg" => "Successfully created news",
                     "status" => 200
-                ]
+                ], 200
             ]);
         } catch (\Throwable $err) {
             //throw $th;
-            return response()->json(["res" => ["msg" => $err->getMessage(), "status" => 400]]);
+            return response()->json(["res" => ["msg" => $err->getMessage(), "status" => 400]], 400);
         }
     }
 }
